Split third party plugins tests into their own classes
This makes it easier to test them individually and benefit
from test distribution.
